+++ ChangeLog.md
@@ -2,7 +2,7 @@
 
 # 3.2.1
 
-  * configure `max_chunk_size` in `InstreamCommand` by `CLAMAV_INSTREAM_MAX_CHUNK_SIZE` env
+  * Add a configuration to custom `max_chunk_size` in `InstreamCommand`
 
 # 3.2.0